[jit] Delete WeakScriptModuleProxy
